\name{inputNetwork}
\alias{inputNetwork}
\docType{data}
\title{
   A network example with graphNEL class
}
\description{
   The network is a sub-network of HPRD (Human Protein Reference Database) network with 320 nodes and 769 edges, which is an example to show how to input network with graphNEL class into the package.
}

\usage{data(inputNetwork)}

\format{graphNEL}

\keyword{datasets}
