\name{netsam_output}
\alias{netsam_output}
\docType{data}
\title{
   An example of the list object returned by NetSAM function
}
\description{
   The list object contains at least three parts: gene order information, module information and network information. This object can be used as an input of the function featureAssociation or GOAssociation.
}
\usage{data(NetSAMOutput_Example)}
\format{list}

\keyword{datasets}
