\name{NetSAM-package}
\alias{NetSAM-package}

\docType{package}
\title{
	Network Seriation and Modularization
}
\description{
	The NetSAM (Network Seriation and Modularization) package takes an edge-list representation of a network as an input, performs network seriation and modularization analysis, and generates as files that can be used as an input for the one-dimensional network visualization tool NetGestalt (http://www.netgestalt.org) or other network analysis. 
}
\details{
	\tabular{ll}{
		Package: \tab NetSAM\cr
		Type: \tab Package\cr
		Version: \tab 0.99.2\cr
		Date: \tab 2013-06-13\cr
		License: \tab LGPL\cr
		LazyLoad: \tab yes\cr
	}
}

\author{
	Jing Wang
	Maintainer: Bing Zhang <bing.zhang@Vanderbilt.edu>
}

\references{
	NetGestalt: integrating multidimensional omics data over biological networks (accepted by Nature Methods).
}

\seealso{
	\code{\link{NetSAM}}
}

\keyword{ package }
