\name{NetSAM-package}
\alias{NetSAM-package}

\docType{package}
\title{
	Network Seriation and Modularization
}
\description{
	The NetSAM (Network Seriation and Modularization) package takes an edge-list representation of a weighted or unweighted network as an input, performs network seriation and modularization analysis, and generates as files that can be used as an input for the one-dimensional network visualization tool NetGestalt (http://www.netgestalt.org) or other network analysis. Meanwhile, the NetSAM package can also generate correlation network (e.g. co-expression network) based on the input matrix data and then perform seriation and modularization analysis for correlation network.
}
\details{
	\tabular{ll}{
		Package: \tab NetSAM\cr
		Type: \tab Package\cr
		Version: \tab 2.0.3\cr
		Date: \tab 2016-11-29\cr
		License: \tab LGPL\cr
		LazyLoad: \tab yes\cr
	}
}

\author{
	Jing Wang
	Maintainer: Jing Wang <jing.wang@bcm.edu>
}

\references{
	NetGestalt: integrating multidimensional omics data over biological networks. Nature Methods 10, 597-598 (2013).
}

\seealso{
	\code{\link{NetSAM}}
	\code{\link{MatSAM}}
}

\keyword{ package }
